Upload FMA_parse.py(not currently working) and FMA_IDs.csv
FMA_parse.py is a program that should take the FMA_ID and parentID information to create nested dictionaries to create a tree for the numbers in the FMA_IDs.csv file. The program is currently not working. Need to find a way to better implement .setdefault() to help create nested dictionary if a certain parent ID is not present.